+++ utils.go
@@ -22,7 +22,6 @@ import (
 
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
-	semconv "go.opentelemetry.io/otel/semconv/v1.24.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -103,7 +102,7 @@ func createSpan(
 ) (context.Context, trace.Span) {
 	attrs := cfg.Attributes
 	if enableDBStatement && !cfg.SpanOptions.DisableQuery {
-		attrs = append(attrs, semconv.DBStatementKey.String(query))
+		attrs = append(attrs, cfg.DBQueryTextAttributes(query)...)
 	}
 	if cfg.AttributesGetter != nil {
 		attrs = append(attrs, cfg.AttributesGetter(ctx, method, query, args)...)